+++ qgreenland/config/layers/Sea ice/Monthly mean concentration (25 km)/Feb or March (max monthly extent)/__settings__.py
@@ -1,0 +1,8 @@
+from qgreenland.config.helpers.layers.sea_ice_concentration import (
+    MAX_CONCENTRATION_YEARS,
+)
+from qgreenland.models.config.layer_group import LayerGroupSettings
+
+settings = LayerGroupSettings(
+    order=[f":seaice_maximum_concentration_{year}" for year in MAX_CONCENTRATION_YEARS],
+)